+++ icu/source/common/ucharstriebuilder.cpp
@@ -1,5 +1,3 @@
-// Copyright (C) 2016 and later: Unicode, Inc. and others.
-// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *   Copyright (C) 2010-2012, International Business Machines
@@ -117,7 +115,7 @@ UCharsTrieBuilder::add(const UnicodeString &s, int32_t value, UErrorCode &errorC
             return *this;
         }
         if(elementsLength>0) {
-            uprv_memcpy(newElements, elements, (size_t)elementsLength*sizeof(UCharsTrieElement));
+            uprv_memcpy(newElements, elements, elementsLength*sizeof(UCharsTrieElement));
         }
         delete[] elements;
         elements=newElements;
