+++ icu/source/test/intltest/simpleformattertest.cpp
@@ -1,5 +1,3 @@
-// Copyright (C) 2016 and later: Unicode, Inc. and others.
-// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 * Copyright (C) 2014-2016, International Business Machines Corporation and
@@ -90,10 +88,10 @@ void SimpleFormatterTest::TestNoArguments() {
 void SimpleFormatterTest::TestSyntaxErrors() {
     UErrorCode status = U_ZERO_ERROR;
     SimpleFormatter fmt("{}", status);
-    assertEquals("syntax error {}", (int32_t)U_ILLEGAL_ARGUMENT_ERROR, status);
+    assertEquals("syntax error {}", U_ILLEGAL_ARGUMENT_ERROR, status);
     status = U_ZERO_ERROR;
     fmt.applyPattern("{12d", status);
-    assertEquals("syntax error {12d", (int32_t)U_ILLEGAL_ARGUMENT_ERROR, status);
+    assertEquals("syntax error {12d", U_ILLEGAL_ARGUMENT_ERROR, status);
 }
 
 void SimpleFormatterTest::TestOneArgument() {
@@ -343,7 +341,7 @@ void SimpleFormatterTest::TestFormatReplaceNoOptimization() {
         return;
     }
     UnicodeString result("original");
-    int32_t offsets[4];
+    int offsets[4];
     UnicodeString freddy("freddy");
     UnicodeString frog("frog");
     UnicodeString by("by");
@@ -373,7 +371,7 @@ void SimpleFormatterTest::TestFormatReplaceNoOptimizationLeadingText() {
         return;
     }
     UnicodeString result("original");
-    int32_t offsets[4];
+    int offsets[4];
     UnicodeString freddy("freddy");
     UnicodeString frog("frog");
     UnicodeString by("by");
@@ -403,7 +401,7 @@ void SimpleFormatterTest::TestFormatReplaceOptimization() {
         return;
     }
     UnicodeString result("original");
-    int32_t offsets[4];
+    int offsets[4];
     UnicodeString freddy("freddy");
     UnicodeString frog("frog");
     UnicodeString by("by");
@@ -433,7 +431,7 @@ void SimpleFormatterTest::TestFormatReplaceNoOptimizationLeadingArgumentUsedTwic
         return;
     }
     UnicodeString result("original");
-    int32_t offsets[4];
+    int offsets[4];
     UnicodeString freddy("freddy");
     UnicodeString frog("frog");
     UnicodeString by("by");
