+++ icu/source/tools/toolutil/swapimpl.cpp
@@ -1,5 +1,3 @@
-// Copyright (C) 2016 and later: Unicode, Inc. and others.
-// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *
@@ -232,7 +230,7 @@ uprops_swap(const UDataSwapper *ds,
 
         /* copy everything for inaccessible data (padding) */
         if(inData32!=outData32) {
-            uprv_memcpy(outData32, inData32, 4*(size_t)dataTop);
+            uprv_memcpy(outData32, inData32, 4*dataTop);
         }
 
         /* swap the indexes[16] */
