+++ icu/source/tools/toolutil/unewdata.c
@@ -1,5 +1,3 @@
-// Copyright (C) 2016 and later: Unicode, Inc. and others.
-// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *
@@ -41,7 +39,7 @@ udata_create(const char *dir, const char *type, const char *name,
     uint16_t headerSize, commentLength;
     char filename[512];
     uint8_t bytes[16];
-    int32_t length;
+    int length;
 
     if(pErrorCode==NULL || U_FAILURE(*pErrorCode)) {
         return NULL;
@@ -76,7 +74,7 @@ udata_create(const char *dir, const char *type, const char *name,
 
         
      /* LDH buffer Length error check */
-    if(length  > ((int32_t)sizeof(filename) - 1))
+    if(length  > (sizeof(filename) - 1))
     {
    	    *pErrorCode = U_BUFFER_OVERFLOW_ERROR;
    	    uprv_free(pData);
